refactor: move code around and optimize
