+++ apps/antalmanac/src/components/Calendar/CalendarToolbar.tsx
@@ -1,42 +1,16 @@
-import {
-    Add as AddIcon,
-    ArrowDropDown as ArrowDropDownIcon,
-    Edit as EditIcon,
-    Undo as UndoIcon,
-    Clear as ClearIcon,
-} from '@mui/icons-material';
-import { Box, Button, IconButton, Paper, Popover, Tooltip, Typography, useTheme } from '@mui/material';
-import { useState, useMemo, useCallback, useEffect } from 'react';
+import { Undo as UndoIcon } from '@mui/icons-material';
+import { Box, Button, IconButton, Paper, Tooltip } from '@mui/material';
+import { useState, useCallback, useEffect, memo } from 'react';
 
 import CustomEventDialog from './Toolbar/CustomEventDialog/CustomEventDialog';
 
-import { changeCurrentSchedule, undoDelete } from '$actions/AppStoreActions';
+import { undoDelete } from '$actions/AppStoreActions';
+import { SelectSchedulePopover } from '$components/Calendar/Toolbar/schedule-select/schedule-select';
 import { ClearScheduleButton } from '$components/buttons/Clear';
-import { CopyScheduleButton } from '$components/buttons/Copy';
 import DownloadButton from '$components/buttons/Download';
 import ScreenshotButton from '$components/buttons/Screenshot';
-import AddScheduleDialog from '$components/dialogs/AddSchedule';
-import DeleteScheduleDialog from '$components/dialogs/DeleteSchedule';
-import RenameScheduleDialog from '$components/dialogs/RenameSchedule';
 import analyticsEnum, { logAnalytics } from '$lib/analytics';
 import AppStore from '$stores/AppStore';
-
-function handleScheduleChange(index: number) {
-    logAnalytics({
-        category: analyticsEnum.calendar.title,
-        action: analyticsEnum.calendar.actions.CHANGE_SCHEDULE,
-    });
-    changeCurrentSchedule(index);
-}
-
-/**
- * Creates an event handler callback that will change the current schedule to the one at a specified index.
- */
-function createScheduleSelector(index: number) {
-    return () => {
-        handleScheduleChange(index);
-    };
-}
 
 function handleUndo() {
     logAnalytics({
@@ -44,222 +18,6 @@ function handleUndo() {
         action: analyticsEnum.calendar.actions.UNDO,
     });
     undoDelete(null);
-}
-
-interface RenameScheduleButtonProps {
-    index: number;
-    disabled?: boolean;
-}
-
-function RenameScheduleButton({ index, disabled }: RenameScheduleButtonProps) {
-    const [open, setOpen] = useState(false);
-
-    const handleOpen = useCallback(() => {
-        setOpen(true);
-    }, []);
-
-    const handleClose = useCallback(() => {
-        setOpen(false);
-    }, []);
-
-    return (
-        <Box>
-            <Tooltip title="Rename Schedule">
-                <span>
-                    <IconButton onClick={handleOpen} size="small" disabled={disabled}>
-                        <EditIcon />
-                    </IconButton>
-                </span>
-            </Tooltip>
-            <RenameScheduleDialog fullWidth open={open} index={index} onClose={handleClose} />
-        </Box>
-    );
-}
-
-interface DeleteScheduleButtonProps {
-    index: number;
-    disabled?: boolean;
-}
-
-function DeleteScheduleButton({ index, disabled }: DeleteScheduleButtonProps) {
-    const [open, setOpen] = useState(false);
-
-    const handleOpen = useCallback(() => {
-        setOpen(true);
-    }, []);
-
-    const handleClose = useCallback(() => {
-        setOpen(false);
-    }, []);
-
-    return (
-        <Box>
-            <Tooltip title="Delete Schedule">
-                <span>
-                    <IconButton
-                        onClick={handleOpen}
-                        size="small"
-                        disabled={AppStore.schedule.getNumberOfSchedules() === 1 || disabled}
-                    >
-                        <ClearIcon />
-                    </IconButton>
-                </span>
-            </Tooltip>
-            <DeleteScheduleDialog fullWidth open={open} index={index} onClose={handleClose} />
-        </Box>
-    );
-}
-
-interface AddScheduleButtonProps {
-    disabled: boolean;
-}
-
-/**
- * MenuItem nested in the select menu to add a new schedule through a dialog.
- */
-function AddScheduleButton({ disabled }: AddScheduleButtonProps) {
-    const [open, setOpen] = useState(false);
-
-    const handleOpen = useCallback(() => {
-        setOpen(true);
-    }, []);
-
-    const handleClose = useCallback(() => {
-        setOpen(false);
-    }, []);
-
-    return (
-        <>
-            <Button color="inherit" onClick={handleOpen} sx={{ display: 'flex', gap: 1 }} disabled={disabled}>
-                <AddIcon />
-                <Typography whiteSpace="nowrap" textOverflow="ellipsis" overflow="hidden" textTransform="none">
-                    Add Schedule
-                </Typography>
-            </Button>
-            <AddScheduleDialog fullWidth open={open} onClose={handleClose} />
-        </>
-    );
-}
-
-/**
- * Simulates an HTML select element using a popover.
- *
- * Can select a schedule, and also control schedule settings with buttons.
- */
-function SelectSchedulePopover(props: { scheduleNames: string[] }) {
-    const [currentScheduleIndex, setCurrentScheduleIndex] = useState(() => AppStore.getCurrentScheduleIndex());
-    const [skeletonMode, setSkeletonMode] = useState(() => AppStore.getSkeletonMode());
-
-    const [anchorEl, setAnchorEl] = useState<HTMLElement>();
-
-    const theme = useTheme();
-
-    // TODO: maybe these widths should be dynamic based on i.e. the viewport width?
-
-    const minWidth = useMemo(() => 100, []);
-    const maxWidth = useMemo(() => 150, []);
-
-    const open = useMemo(() => Boolean(anchorEl), [anchorEl]);
-
-    const currentScheduleName = useMemo(() => {
-        return props.scheduleNames[currentScheduleIndex];
-    }, [props.scheduleNames, currentScheduleIndex]);
-
-    const handleClick = useCallback((event: React.MouseEvent<HTMLButtonElement>) => {
-        setAnchorEl(event.currentTarget);
-    }, []);
-
-    const handleClose = useCallback(() => {
-        setAnchorEl(undefined);
-    }, []);
-
-    const handleScheduleIndexChange = useCallback(() => {
-        setCurrentScheduleIndex(AppStore.getCurrentScheduleIndex());
-    }, []);
-
-    const handleSkeletonModeChange = () => {
-        setSkeletonMode(AppStore.getSkeletonMode());
-    };
-
-    useEffect(() => {
-        AppStore.on('addedCoursesChange', handleScheduleIndexChange);
-        AppStore.on('customEventsChange', handleScheduleIndexChange);
-        AppStore.on('colorChange', handleScheduleIndexChange);
-        AppStore.on('currentScheduleIndexChange', handleScheduleIndexChange);
-        AppStore.on('skeletonModeChange', handleSkeletonModeChange);
-
-        return () => {
-            AppStore.off('addedCoursesChange', handleScheduleIndexChange);
-            AppStore.off('customEventsChange', handleScheduleIndexChange);
-            AppStore.off('colorChange', handleScheduleIndexChange);
-            AppStore.off('currentScheduleIndexChange', handleScheduleIndexChange);
-            AppStore.off('skeletonModeChange', handleSkeletonModeChange);
-        };
-    }, [handleScheduleIndexChange]);
-
-    return (
-        <Box>
-            <Button
-                size="small"
-                color="inherit"
-                variant="outlined"
-                onClick={handleClick}
-                sx={{ minWidth, maxWidth, justifyContent: 'space-between' }}
-            >
-                <Typography whiteSpace="nowrap" textOverflow="ellipsis" overflow="hidden" textTransform="none">
-                    {currentScheduleName}
-                </Typography>
-                <ArrowDropDownIcon />
-            </Button>
-
-            <Popover
-                open={open}
-                anchorEl={anchorEl}
-                onClose={handleClose}
-                anchorOrigin={{ vertical: 'bottom', horizontal: 'left' }}
-            >
-                <Box padding={1}>
-                    {props.scheduleNames.map((name, index) => (
-                        <Box key={index} display="flex" alignItems="center" gap={1}>
-                            <Box flexGrow={1}>
-                                <Button
-                                    color="inherit"
-                                    sx={{
-                                        minWidth,
-                                        maxWidth,
-                                        width: '100%',
-                                        display: 'flex',
-                                        justifyContent: 'flex-start',
-                                        background:
-                                            index === currentScheduleIndex ? theme.palette.action.selected : undefined,
-                                    }}
-                                    onClick={createScheduleSelector(index)}
-                                >
-                                    <Typography
-                                        overflow="hidden"
-                                        whiteSpace="nowrap"
-                                        textTransform="none"
-                                        textOverflow="ellipsis"
-                                    >
-                                        {name}
-                                    </Typography>
-                                </Button>
-                            </Box>
-                            <Box display="flex" alignItems="center" gap={0.5}>
-                                <CopyScheduleButton index={index} disabled={skeletonMode} />
-                                <RenameScheduleButton index={index} disabled={skeletonMode} />
-                                <DeleteScheduleButton index={index} disabled={skeletonMode} />
-                            </Box>
-                        </Box>
-                    ))}
-
-                    <Box marginY={1} />
-
-                    <AddScheduleButton disabled={skeletonMode} />
-                </Box>
-            </Popover>
-        </Box>
-    );
 }
 
 export interface CalendarPaneToolbarProps {
@@ -272,7 +30,7 @@ export interface CalendarPaneToolbarProps {
 /**
  * The root toolbar will pass down the schedule names to its children.
  */
-function CalendarPaneToolbar(props: CalendarPaneToolbarProps) {
+export const CalendarToolbar = memo((props: CalendarPaneToolbarProps) => {
     const { showFinalsSchedule, toggleDisplayFinalsSchedule } = props;
     const [scheduleNames, setScheduleNames] = useState(AppStore.getScheduleNames());
     const [skeletonMode, setSkeletonMode] = useState(AppStore.getSkeletonMode());
@@ -359,6 +117,6 @@ function CalendarPaneToolbar(props: CalendarPaneToolbarProps) {
             </Box>
         </Paper>
     );
-}
+});
 
-export default CalendarPaneToolbar;
+CalendarToolbar.displayName = 'CalendarToolbar';
